requirements.txt
bs4
requests
python-telegram
python-telegram-vot
django
django-extensions